♻️ Refactor keyboard functionality into separate module
