+++ src-tauri/src/lib.rs
@@ -1,11 +1,10 @@
-use enigo::Keyboard;
 use serde::Serialize;
 use tauri::Emitter;
-use tauri_plugin_clipboard_manager::ClipboardExt;
 use tauri_plugin_notification::NotificationExt;
 use std::sync::{Mutex, OnceLock};
 
 mod cli;
+mod keyboard;
 
 #[cfg(target_os = "macos")]
 use macos_accessibility_client;
@@ -15,42 +14,6 @@ use macos_accessibility_client;
 async fn get_selected_text() -> Result<String, String> {
     let text = get_selected_text::get_selected_text().map_err(|e| e.to_string())?;
     Ok(text)
-}
-
-fn select_all_sync() -> Result<(), String> {
-    let mut enigo = enigo::Enigo::new(&enigo::Settings::default()).map_err(|e| e.to_string())?;
-    let modifier = if cfg!(target_os = "macos") {
-        enigo::Key::Meta
-    } else {
-        enigo::Key::Control
-    };
-    enigo.key(modifier, enigo::Direction::Press).map_err(|e| e.to_string())?;
-    enigo.key(enigo::Key::Unicode('a'), enigo::Direction::Click).map_err(|e| e.to_string())?;
-    enigo.key(modifier, enigo::Direction::Release).map_err(|e| e.to_string())?;
-    Ok(())
-}
-
-#[tauri::command]
-async fn select_all(_app: tauri::AppHandle) -> Result<(), String> {
-    #[cfg(target_os = "macos")]
-    {
-        let (tx, rx) = std::sync::mpsc::sync_channel::<Result<(), String>>(1);
-        _app
-            .run_on_main_thread(move || {
-                let _ = tx.send(select_all_sync());
-            })
-            .map_err(|e| e.to_string())?;
-        tauri::async_runtime::spawn_blocking(move || match rx.recv() {
-            Ok(r) => r,
-            Err(_) => Err("select_all was cancelled".to_string()),
-        })
-        .await
-        .map_err(|e| e.to_string())?
-    }
-    #[cfg(not(target_os = "macos"))]
-    {
-        select_all_sync()
-    }
 }
 
 #[tauri::command]
@@ -174,74 +137,6 @@ fn request_mac_accessibility_permissions() -> Result<bool, String> {
     }
 }
 
-fn type_text_sync(text: String, window: tauri::Window) -> Result<(), String> {
-    let mut enigo = enigo::Enigo::new(&enigo::Settings::default()).map_err(|e| e.to_string())?;
-
-    // 1. Save current clipboard content to restore later
-    let previous_clipboard = window.clipboard().read_text().unwrap_or_default();
-
-    // 2. Write the new text to the clipboard
-    window
-        .clipboard()
-        .write_text(text.clone())
-        .map_err(|e| e.to_string())?;
-
-    // Small delay to ensure clipboard is updated
-    std::thread::sleep(std::time::Duration::from_millis(50));
-
-    // 3. Ctrl/Cmd + V to paste text
-    let control_or_command = if cfg!(target_os = "macos") {
-        enigo::Key::Meta
-    } else {
-        enigo::Key::Control
-    };
-    enigo
-        .key(control_or_command, enigo::Direction::Press)
-        .map_err(|e| e.to_string())?;
-    enigo
-        .key(enigo::Key::Unicode('v'), enigo::Direction::Click)
-        .map_err(|e| e.to_string())?;
-    enigo
-        .key(control_or_command, enigo::Direction::Release)
-        .map_err(|e| e.to_string())?;
-
-    // 4. Small delay to ensure paste is completed before restoring clipboard
-    std::thread::sleep(std::time::Duration::from_millis(50));
-
-    // 5. Restore previous clipboard
-    if !previous_clipboard.is_empty() {
-        let _ = window.clipboard().write_text(previous_clipboard);
-    }
-
-    Ok(())
-}
-
-#[tauri::command]
-async fn type_text(text: String, window: tauri::Window) -> Result<(), String> {
-    #[cfg(target_os = "macos")]
-    {
-        let (tx, rx) = std::sync::mpsc::sync_channel::<Result<(), String>>(1);
-        let window_for_task = window.clone();
-        window
-            .run_on_main_thread(move || {
-                let _ = tx.send(type_text_sync(text, window_for_task));
-            })
-            .map_err(|e| e.to_string())?;
-        tauri::async_runtime::spawn_blocking(move || match rx.recv() {
-            Ok(r) => r,
-            Err(_) => Err("type_text was cancelled".to_string()),
-        })
-        .await
-        .map_err(|e| e.to_string())?
-    }
-    #[cfg(not(target_os = "macos"))]
-    {
-        tauri::async_runtime::spawn_blocking(move || type_text_sync(text, window))
-            .await
-            .map_err(|e| e.to_string())?
-    }
-}
-
 #[cfg_attr(mobile, tauri::mobile_entry_point)]
 pub fn run() {
     let startup_selection = cli::has_selection_flag(std::env::args());
@@ -275,8 +170,8 @@ pub fn run() {
         })
         .invoke_handler(tauri::generate_handler![
             get_selected_text,
-            select_all,
-            type_text,
+            keyboard::select_all,
+            keyboard::type_text,
             get_platform_info,
             get_session_info,
             consume_pending_selection_input,
